requirements.txt
discord.py>=2.3.0
playwright>=1.40.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiohttp>=3.9.0
Pillow>=10.0.0
asyncio-throttle>=1.0.0
